refactor: merge Abstract with LeaderAppender
The AbstractAppender base class added unnecessary complexity to the LeaderAppender. Removing the base class simplifies the code and makes certain things clearer.
